requirements.txt
APScheduler==3.0.0
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
bundle==1.1.2
celery==3.1.15
docopt==0.6.2
feedparser==5.1.3
itsdangerous==0.24
kombu==3.0.23
pymongo==2.7.2
pytz==2014.7
six==1.8.0
tzlocal==1.1.1
yolk==0.4.3
